@@ -4,26 +4,25 @@
    "version": "1.0.0",
    "private": true,
    "dependencies": {
-      "@reduxjs/toolkit": "^1.6.1",
-      "axios": "^0.27.2",
-      "bootstrap": "^5.1.3",
+      "@reduxjs/toolkit": "^1.9.1",
+      "axios": "^1.2.0",
+      "bootstrap": "^5.2.3",
       "colors-generator": "^0.3.4",
       "colorsys": "^1.0.22",
-      "detect-file-type": "^0.2.8",
-      "filesize": "^9.0.0",
+      "filesize": "^10.0.5",
       "idb-keyval": "^6.2.0",
       "lodash.orderby": "^4.6.0",
-      "ol": "^6.14.1",
+      "ol": "^7.1.0",
       "proj4": "^2.8.0",
-      "react": "^17.0.2",
-      "react-bootstrap": "^2.4.0",
-      "react-dom": "^17.0.2",
-      "react-redux": "^7.2.6",
-      "react-scripts": "4.0.3",
-      "react-sortablejs": "^6.1.1",
+      "react": "^18.2.0",
+      "react-bootstrap": "^2.6.0",
+      "react-dom": "^18.2.0",
+      "react-redux": "^8.0.5",
+      "react-scripts": "5.0.1",
+      "react-sortablejs": "^6.1.4",
       "react-split": "^2.0.14",
       "redux": "^4.2.0",
-      "sass": "^1.52.1",
+      "sass": "^1.56.1",
       "sortablejs": "^1.15.0",
       "url-parse": "^1.5.10"
    },
@@ -51,5 +50,8 @@
          "last 1 firefox version",
          "last 1 safari version"
       ]
+   },
+   "resolutions": {
+      "react-error-overlay": "6.0.11"
    }
 }
