@@ -1,6 +1,6 @@
 import { GeometryCollection } from 'ol/geom';
 import { Stroke, Style, Fill, Circle } from 'ol/style';
-import GeometryType from 'ol/geom/GeometryType';
+import { GeometryType } from './constants';
 import { getAreaFormatted, getFeaturesByName, getLayer, getLengthFormatted, groupBy } from './helpers';
 import WKT from 'ol/format/WKT';
 import { filterSelector } from 'utils/sld-reader/Filter';
@@ -37,14 +37,14 @@ export function addGeometryInfo(features) {
       const geometryType = geometry.getType();
 
       switch (geometryType) {
-         case GeometryType.POLYGON:
-         case GeometryType.MULTI_POLYGON:
+         case GeometryType.Polygon:
+         case GeometryType.MultiPolygon:
             if (!feature.get('_area')) {
                feature.set('_area', getAreaFormatted(geometry));
             }
             break;
-         case GeometryType.LINE_STRING:
-         case GeometryType.MULTI_LINE_STRING:
+         case GeometryType.LineString:
+         case GeometryType.MultiLineString:
             if (!feature.get('_length')) {
                feature.set('_length', getLengthFormatted(geometry));
             }
@@ -188,7 +188,7 @@ function getHighlightStyle(feature) {
 
          highlightStyle = styles[0].clone();
          highlightStyle.getText().setStroke(stroke);
-      } else if (feature.getGeometry().getType() === GeometryType.POINT) {
+      } else if (feature.getGeometry().getType() === GeometryType.Point) {
          const image = feature.getStyle()[0].getImage();
 
          highlightStyle = new Style({
@@ -222,13 +222,13 @@ function addZoomToStyle(geometries, styles) {
       const geometryType = geometry.getType();
 
       switch (geometryType) {
-         case GeometryType.GEOMETRY_COLLECTION:
+         case GeometryType.GeometryCollection:
             addZoomToStyle(geometry.getGeometries(), styles);
             break;
-         case GeometryType.POLYGON:
-         case GeometryType.MULTI_POLYGON:
-         case GeometryType.LINE_STRING:
-         case GeometryType.MULTI_LINE_STRING:
+         case GeometryType.Polygon:
+         case GeometryType.MultiPolygon:
+         case GeometryType.LineString:
+         case GeometryType.MultiLineString:
             styles.push(
                new Style({
                   geometry,
@@ -240,8 +240,8 @@ function addZoomToStyle(geometries, styles) {
                })
             );
             break;
-         case GeometryType.POINT:
-         case GeometryType.MULTI_POINT:
+         case GeometryType.Point:
+         case GeometryType.MultiPoint:
             styles.push(
                new Style({
                   geometry,
